Use screen_block_draw to clear prefs area
No need to do all of that scaling ourselves when we can just use the
screen routines to do it for us.

Since screen_block_draw asserts flow control, it was necessary
to add a dummy flow-control function, that the flow on/off function
pointers can point at until the drivers are loaded.